@@ -1,6 +1,5 @@
 use anchor_lang::prelude::*;
-use anchor_spl::token;
-use anchor_spl::token::{Token, MintTo, Transfer};
+use anchor_spl::token::{Mint, MintTo, Token, Transfer};
 
 
 declare_id!("Fg6PaFpoGXkYsidMpWTK6W2BeZ7FEfcYkg476zPFsLnS");
@@ -9,42 +8,41 @@ declare_id!("Fg6PaFpoGXkYsidMpWTK6W2BeZ7FEfcYkg476zPFsLnS");
 pub mod token_contract {
     use super::*;
 
-    pub fn mint_token(ctx: Context<MintToken>,) -> Result<()> {
+    pub fn mint_token(ctx: Context<MintToken>) -> Result<()> {
         // Create the MintTo struct for our context
         let cpi_accounts = MintTo {
             mint: ctx.accounts.mint.to_account_info(),
             to: ctx.accounts.token_account.to_account_info(),
             authority: ctx.accounts.authority.to_account_info(),
         };
-        
+
         let cpi_program = ctx.accounts.token_program.to_account_info();
         // Create the CpiContext we need for the request
         let cpi_ctx = CpiContext::new(cpi_program, cpi_accounts);
 
         // Execute anchor's helper function to mint tokens
-        token::mint_to(cpi_ctx, 10)?;
-        
+        anchor_spl::token::mint_to(cpi_ctx, 10)?;
+
         Ok(())
     }
 
     pub fn transfer_token(ctx: Context<TransferToken>) -> Result<()> {
         // Create the Transfer struct for our context
-        let transfer_instruction = Transfer{
+        let transfer_instruction = Transfer {
             from: ctx.accounts.from.to_account_info(),
             to: ctx.accounts.to.to_account_info(),
             authority: ctx.accounts.from_authority.to_account_info(),
         };
-         
+
         let cpi_program = ctx.accounts.token_program.to_account_info();
         // Create the Context for our Transfer request
         let cpi_ctx = CpiContext::new(cpi_program, transfer_instruction);
 
         // Execute anchor's helper function to transfer tokens
         anchor_spl::token::transfer(cpi_ctx, 5)?;
- 
+
         Ok(())
     }
-
 }
 
 #[derive(Accounts)]
